git-svn-id: svn+ssh://mercurium.pc.ac.upc.edu/var/lib/svn/mcxx/trunk@373 39d7fdcd-100c-0410-a620-8454655e6f6e

+++ src/cxx-macros.h
@@ -35,6 +35,6 @@
   #define MCXX_END_DECLS
 #endif
 
-#define BITMAP_TEST(x, b) (((x) & (b)) == (b))
+#define BITMAP_TEST(x, b) ((((int)(x)) & ((int)(b))) == (b))
 
 #endif // CXX_MACROS_H
